updated both frontend & backend to match each others implementations
