+++ server/app/controllers/jobs_controller.py
@@ -1,6 +1,5 @@
 from sqlalchemy.orm import Session
-from app.schemas.jobs_schema import Job, JobDTO
-from app.schemas.suburbs_schema import Suburb
+from app.schemas.jobs_schema import JobDTO
 from app.models.jobs_model import Jobs 
 from typing import List
 
@@ -37,10 +36,13 @@ def get_all_accepted_jobs(db: Session) -> List[JobDTO]:
 controller to handle updating the job status to either 'accepted' or 'declined
 """
 def update_job_status(db: Session, id: int, status: str) -> None:
+    print("cool!")
     if(not Jobs.does_job_exist(db, id)):
         return {
             "Error": f"Job: {id} does not exist"
         }
+
+    print("yoyo")
 
     Jobs.update_job_status(db, id, status)
