+++ server/app/models/suburb_model.py
@@ -1,8 +1,6 @@
 from app.db.core import Base
-from sqlalchemy import Boolean, VARCHAR, DateTime, INTEGER, TEXT, TIMESTAMP, ForeignKey
-from sqlalchemy.dialects.postgresql import UUID
+from sqlalchemy import VARCHAR, INTEGER
 from sqlalchemy.sql.schema import Column
-from sqlalchemy.orm import Session, relationship
 from app.schemas.suburbs_schema import Suburb
 
 class Suburb(Base):
