+++ server/app/schemas/suburbs_schema.py
@@ -1,6 +1,4 @@
-from typing import Optional
 from pydantic import BaseModel
-from datetime import datetime 
 
 class Suburb(BaseModel):
     id: int
