--- ui/src/JobOptions/constants.ts
@@ -1,2 +1,0 @@
-export const INVITED_JOB_OPTION = 'Invited';
-export const ACCEPTED_JOB_OPTION = 'Accepted';
