+++ ui/src/Jobs/JobInvitations/JobInvitation.tsx
@@ -1,16 +1,19 @@
 import React from 'react';
-import { JobContactInfo } from '../commonComponents/JobUI/JobContactInfo';
-import { JobDescription } from '../commonComponents/JobUI/JobDescription';
-import { JobLocationInfo } from '../commonComponents/JobUI/JobLocationInfo';
-import { Button } from '../commonComponents/buttons/Button';
-import { Job } from '../types';
+import { JobContactInfo } from '../../commonComponents/JobUI/JobContactInfo';
+import { JobDescription } from '../../commonComponents/JobUI/JobDescription';
+import { JobLocationInfo } from '../../commonComponents/JobUI/JobLocationInfo';
+import { Button } from '../../commonComponents/buttons/Button';
+import { Job, JobStatusKeys } from '../../types';
+import { ACCEPTED_JOB_STATUS, DECLINED_JOB_STATUS } from '../../constants';
 
 interface Props {
     jobDetails: Job
+    onHandleJob: (id: number, status: JobStatusKeys) => void;
 }
 
 export const JobInvitation:React.FC<Props> = ({ 
-    jobDetails
+    jobDetails,
+    onHandleJob
 }) => {
     const {
         contact_name,
@@ -39,9 +42,10 @@ export const JobInvitation:React.FC<Props> = ({
             />
 
             <section className='button-group'>
-                <Button type='primary'>Accept</Button>
-                <Button type='secondary'>Decline</Button>
+                <Button type='primary' onClick={() => onHandleJob(id, ACCEPTED_JOB_STATUS)}>Accept</Button>
+                <Button type='secondary' onClick={() => onHandleJob(id, DECLINED_JOB_STATUS)}>Decline</Button>
                 <p>${price}</p>
+                {/* //handle specific error here */}
             </section>
         </section>
     )
