+++ ui/src/Jobs/JobInvitations/jobInvitationStyles.css
