+++ ui/src/Jobs/JobOptions/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { JobOptionKeys } from '../types';
+import { JOB_OPTION_ACCEPTED_KEY, JOB_OPTION_INVITED_KEY } from '../../constants';
+import { JobOptionKeys } from '../../types';
 import { OptionTab } from './components/OptionTab';
-import { ACCEPTED_JOB_OPTION, INVITED_JOB_OPTION } from './constants';
 import './jobOptionStyles.css';
 
 interface Props {
@@ -15,14 +15,14 @@ export const JobOptions:React.FC<Props> = ({ children, onChangeOption, optionKey
         <>
             <section className='job-options-container'>
                 <OptionTab 
-                selected={optionKey === INVITED_JOB_OPTION} 
-                optionName={INVITED_JOB_OPTION} 
-                onClick={() => onChangeOption(INVITED_JOB_OPTION)}
+                selected={optionKey === JOB_OPTION_INVITED_KEY} 
+                optionName={JOB_OPTION_INVITED_KEY} 
+                onClick={() => onChangeOption(JOB_OPTION_INVITED_KEY)}
                 />
                 <OptionTab 
-                   selected={optionKey === ACCEPTED_JOB_OPTION} 
-                   optionName={ACCEPTED_JOB_OPTION} 
-                   onClick={() => onChangeOption(ACCEPTED_JOB_OPTION)}
+                   selected={optionKey === JOB_OPTION_ACCEPTED_KEY} 
+                   optionName={JOB_OPTION_ACCEPTED_KEY} 
+                   onClick={() => onChangeOption(JOB_OPTION_ACCEPTED_KEY)}
                 />
             </section>
             {children}
