+++ ui/src/constants.ts
@@ -1,0 +1,5 @@
+export const JOB_OPTION_INVITED_KEY = 'Invited';
+export const JOB_OPTION_ACCEPTED_KEY = 'Accepted';
+
+export const ACCEPTED_JOB_STATUS = 'accepted';
+export const DECLINED_JOB_STATUS = 'declined';
